# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ResponseData_any_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"items": {},
						"type": "array"
					},
					"data": {
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ResponseData_any-Array_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"items": {},
						"type": "array"
					},
					"data": {
						"items": {},
						"type": "array",
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CustomerErrorResponse": {
				"description": "Specialized response for Customer errors to show possible messages in Swagger",
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"$ref": "#/components/schemas/CustomerErrorMessage"
					},
					"error": {
						"items": {},
						"type": "array"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseData_CustomerProfileResponse-or-null_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"items": {},
						"type": "array"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/CustomerProfileResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ResponseData_string-or-any_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"items": {},
						"type": "array"
					},
					"data": {
						"anyOf": [
							{
								"type": "string"
							},
							{}
						],
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ResponseData_DashboardStatsResponse_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"items": {},
						"type": "array"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/DashboardStatsResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"AdminErrorResponse": {
				"description": "Specialized response for Admin errors to show possible messages in Swagger",
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"$ref": "#/components/schemas/AdminErrorMessage",
						"description": "Admin error code. Possible values:\n- MODERATOR_NOT_FOUND: Moderator ID does not exist\n- USER_NOT_FOUND: General user not found\n- ACCOUNT_NOT_FOUND: Linked account not found\n- POST_NOT_FOUND: Target post for moderation not found\n- REQUIRED_REASON: Reason is required when rejecting a post\n- REJECTION_REASON_REQUIRED: Specific validation error for rejection\n- PHONE_ALREADY_EXISTS: Phone number already in use\n- EMAIL_ALREADY_EXISTS: Email already in use\n- INTERNAL_SERVER_ERROR: unexpected server error"
					},
					"error": {
						"items": {},
						"type": "array"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"status",
					"message",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ModeratePostRequest": {
				"properties": {
					"actionType": {
						"type": "string",
						"enum": [
							"Approved",
							"Rejected"
						],
						"description": "Action to perform on the post",
						"example": "Approved",
						"pattern": "^(Approved|Rejected)$"
					},
					"reason": {
						"type": "string",
						"description": "Reason for rejection (required if actionType is Rejected)",
						"minLength": 1
					},
					"isHateContent": {
						"type": "boolean",
						"description": "Whether the content is considered hate speech"
					}
				},
				"required": [
					"actionType"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "tro-tot-vn-be",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "lamdx4"
		}
	},
	"paths": {
...
		"/customer/me": {
			"get": {
				"operationId": "GetMyProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_CustomerProfileResponse-or-null_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustomerErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustomerErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get private profile of the current authenticated user",
				"tags": [
					"Customer"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "UpdateMyProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					}
				},
				"description": "Update the current user's profile",
				"tags": [
					"Customer"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"avatar": {
										"type": "string",
										"format": "binary"
									},
									"firstName": {
										"type": "string"
									},
									"lastName": {
										"type": "string"
									},
									"gender": {
										"type": "string"
									},
									"birthday": {
										"type": "string"
									},
									"currentCity": {
										"type": "string"
									},
									"currentDistrict": {
										"type": "string"
									},
									"currentJob": {
										"type": "string"
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/admin/dashboard-stats": {
			"get": {
				"operationId": "GetStatisticsForDashBoard",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_DashboardStatsResponse_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get overall dashboard statistics.",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"Admin",
							"Manager"
						]
					}
				],
				"parameters": []
			}
		},
		"/admin/moderators/{moderatorId}/reset-password": {
			"put": {
				"operationId": "ResetPasswordOfModerator",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_string-or-any_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminErrorResponse"
								}
							}
						}
					}
				},
				"description": "Reset moderator password.",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"Manager"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "moderatorId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/admin/posts/pending": {
			"get": {
				"operationId": "ListPostPending",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any-Array_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					}
				},
				"description": "List all pending posts for review.",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"Admin",
							"Manager"
						]
					}
				],
				"parameters": []
			}
		},
		"/admin/posts/{postId}/moderate": {
			"post": {
				"operationId": "ModeratePost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminErrorResponse"
								}
							}
						}
					}
				},
				"description": "Moderate a post (Approve or Reject).",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"Admin",
							"Manager"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "postId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ModeratePostRequest"
							}
						}
					}
				}
			}
		},
		"/admin/posts/{postId}/history": {
			"get": {
				"operationId": "GetHistoryOfPost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_PostModerationHistoryResponse-Array_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get moderation history for a specific post.",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"Admin",
							"Manager"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "postId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/admin/moderators/{moderatorId}/history": {
			"get": {
				"operationId": "GetHistoryByModeratorId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_ModeratorActionHistoryResponse-Array_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get moderation history for a specific moderator.",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"Manager"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "moderatorId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/admin/moderators": {
			"get": {
				"operationId": "GetModerators",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_ModeratorSummary-Array_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					}
				},
				"description": "List all moderators (excluding managers).",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"Manager"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "key",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "AddModerators",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_string-or-any_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					}
				},
				"description": "Create a new moderator account.",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"Manager"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddModeratorRequest"
							}
						}
					}
				}
			}
		},
		"/admin/moderators/{moderatorId}/status": {
			"put": {
				"operationId": "UpdateModerator",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_string-or-any_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminErrorResponse"
								}
							}
						}
					}
				},
				"description": "Update moderator account status.",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"Manager"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "moderatorId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateModeratorStatusRequest"
							}
						}
					}
				}
			}
		},
		"/admin/moderators/{moderatorId}/profile": {
			"get": {
				"operationId": "GetProfileModerator",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_ModeratorProfileResponse_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get detailed profile of a moderator.",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"Manager"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "moderatorId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/admin/me/profile": {
			"get": {
				"operationId": "GetMyProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					}
				},
				"description": "Get current moderator's own profile.",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"Admin",
							"Manager"
						]
					}
				],
				"parameters": []
			},
			"put": {
				"operationId": "UpdateMyProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_string-or-any_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					}
				},
				"description": "Update current moderator's own profile.",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"Admin",
							"Manager"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/admin/moderation-stats": {
			"get": {
				"operationId": "GetModerationStats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_ModerationStatsResponse_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					}
				},
				"description": "Get moderation performance statistics.",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"Admin",
							"Manager"
						]
					}
				],
				"parameters": []
			}
		},
		"/admin/export-training-data": {
			"get": {
				"operationId": "ExportTrainingData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData_any_"
								}
							}
						}
					}
				},
				"description": "Export training data for AI retraining.",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"Admin",
							"Manager"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}